feat!: reconcile SDK resources in-process and drop Terraform CLI from image (#65)
* feat!: reconcile SDK resources in-process and drop Terraform CLI from image

Move all 106 Terraform Plugin SDK resources from upjet's CLI-based
workspace mode to the in-process SDKv2 architecture
(WithTerraformPluginSDKIncludeList + WithTerraformProvider). All
resources now reconcile natively against the compiled-in provider, so
the container image no longer ships the Terraform CLI (1.8.1, flagged
by security scans and BUSL-licensed) or the mirrored native provider
binary — only the provider Go binary remains.

- config: pass ovh.Provider() for in-process SDKv2 reconciliation and
  the (empty) CLIReconciledResourceList to override upjet's catch-all
  default include list
- controllers: regenerate; 196 controllers switch from
  tjcontroller.NewConnector to NewTerraformPluginSDKAsyncConnector
- image: remove terraform zip, provider filesystem mirror,
  terraformrc.hcl and related build args; bump base image from
  alpine 3.17.1 (EOL) to alpine 3.22
- docs: update AGENTS.md, README and the upgrade skill; the Terraform
  CLI is still used at build time to generate config/schema.json but
  never ships in the image

BREAKING CHANGE: CRD schemas are now generated from the compiled-in Go
provider schema instead of the JSON schema export. 506 numeric fields
across 86 CRDs change from type: number to type: integer
(format: int64), and spec.forProvider.projectId on the vrack
Cloudproject kind is now enforced as required.

Closes #64


* fix: configure SDKv2 provider meta for in-process reconciliation

TerraformSetupBuilder never set terraform.Setup.Meta. In CLI mode that was
harmless, but on the in-process SDKv2 path upjet passes ts.Meta straight
through as the provider meta to schemaMap.Diff, RefreshWithoutUpgrade and
Apply, and never populates it itself -- unlike the Framework path, where it
builds a DynamicValue from ts.Configuration and calls ConfigureProvider.

Every OVH SDKv2 resource begins with meta.(*Config), so a nil Meta panicked
on the first Observe of all 106 SDK-backed resources. The panic is recovered
by controller-runtime and by the async goroutines, so it surfaced as those
resources erroring forever rather than as a crashloop.

Configure an ovh.Provider() from the resolved provider configuration and hand
its Meta() to upjet. ovh.ConfigureContextFunc runs Config.loadAndValidate,
which calls GET /auth/details, so configured metas are cached to keep that to
one call instead of one per Connect. Cache entries are keyed by ProviderConfig
identity plus a hash of the effective configuration, so a rotated OAuth access
token yields a new key and a cached meta is never reused with credentials that
have since changed. A sync.Once per entry collapses a cold-start burst into a
single configuration call, failures are not latched, and idle entries are
evicted after an hour.

Also document this release's CRD schema changes in the release notes: the
number -> integer correction, the x-kubernetes-map-type: granular removal on
User.openstackRc and PrivateNetwork.regionsOpenstackIds, and the new
CloudProject projectId validation rule.

---------

+++ apis/cluster/databases/v1alpha1/zz_projectdatabasepostgresqlconnectionpool_types.go
@@ -45,7 +45,7 @@ type ProjectDatabasePostgresqlConnectionPoolInitParameters struct {
 
 	// Size of the connection pool.
 	// Size of the connection pool
-	Size *float64 `json:"size,omitempty" tf:"size,omitempty"`
+	Size *int64 `json:"size,omitempty" tf:"size,omitempty"`
 
 	// Database user authorized to connect to the pool, if none all the users are allowed.
 	// Database user authorized to connect to the pool, if none all the users are allowed
@@ -74,7 +74,7 @@ type ProjectDatabasePostgresqlConnectionPoolObservation struct {
 
 	// Port of the connection pool.
 	// Port of the connection pool
-	Port *float64 `json:"port,omitempty" tf:"port,omitempty"`
+	Port *int64 `json:"port,omitempty" tf:"port,omitempty"`
 
 	// Ssl connection mode for the pool.
 	// SSL connection mode for the pool
@@ -85,7 +85,7 @@ type ProjectDatabasePostgresqlConnectionPoolObservation struct {
 
 	// Size of the connection pool.
 	// Size of the connection pool
-	Size *float64 `json:"size,omitempty" tf:"size,omitempty"`
+	Size *int64 `json:"size,omitempty" tf:"size,omitempty"`
 
 	// Connection URI to the pool.
 	// Connection URI to the pool
@@ -134,7 +134,7 @@ type ProjectDatabasePostgresqlConnectionPoolParameters struct {
 	// Size of the connection pool.
 	// Size of the connection pool
 	// +kubebuilder:validation:Optional
-	Size *float64 `json:"size,omitempty" tf:"size,omitempty"`
+	Size *int64 `json:"size,omitempty" tf:"size,omitempty"`
 
 	// Database user authorized to connect to the pool, if none all the users are allowed.
 	// Database user authorized to connect to the pool, if none all the users are allowed
