+++ apis/cluster/dns/v1alpha1/zz_dsrecords_types.go
@@ -29,7 +29,7 @@ type DsRecordsDsRecordsInitParameters struct {
 
 	// The record tag
 	// Tag of the DNSSEC key
-	Tag *float64 `json:"tag,omitempty" tf:"tag,omitempty"`
+	Tag *int64 `json:"tag,omitempty" tf:"tag,omitempty"`
 }
 
 type DsRecordsDsRecordsObservation struct {
@@ -48,7 +48,7 @@ type DsRecordsDsRecordsObservation struct {
 
 	// The record tag
 	// Tag of the DNSSEC key
-	Tag *float64 `json:"tag,omitempty" tf:"tag,omitempty"`
+	Tag *int64 `json:"tag,omitempty" tf:"tag,omitempty"`
 }
 
 type DsRecordsDsRecordsParameters struct {
@@ -71,7 +71,7 @@ type DsRecordsDsRecordsParameters struct {
 	// The record tag
 	// Tag of the DNSSEC key
 	// +kubebuilder:validation:Optional
-	Tag *float64 `json:"tag" tf:"tag,omitempty"`
+	Tag *int64 `json:"tag" tf:"tag,omitempty"`
 }
 
 type DsRecordsInitParameters struct {
