+++ apis/cluster/lb/v1alpha1/zz_tcpfrontend_types.go
@@ -24,10 +24,10 @@ type TCPFrontendInitParameters struct {
 	DedicatedIpfo []*string `json:"dedicatedIpfo,omitempty" tf:"dedicated_ipfo,omitempty"`
 
 	// Default TCP Farm of your frontend
-	DefaultFarmID *float64 `json:"defaultFarmId,omitempty" tf:"default_farm_id,omitempty"`
+	DefaultFarmID *int64 `json:"defaultFarmId,omitempty" tf:"default_farm_id,omitempty"`
 
 	// Default ssl served to your customer
-	DefaultSSLID *float64 `json:"defaultSslId,omitempty" tf:"default_ssl_id,omitempty"`
+	DefaultSSLID *int64 `json:"defaultSslId,omitempty" tf:"default_ssl_id,omitempty"`
 
 	// Deny IP Load Balancing access to these ip block. No restriction if null. You cannot specify both allowed_source and denied_source at the same time. List of IP blocks.
 	// +listType=set
@@ -63,10 +63,10 @@ type TCPFrontendObservation struct {
 	DedicatedIpfo []*string `json:"dedicatedIpfo,omitempty" tf:"dedicated_ipfo,omitempty"`
 
 	// Default TCP Farm of your frontend
-	DefaultFarmID *float64 `json:"defaultFarmId,omitempty" tf:"default_farm_id,omitempty"`
+	DefaultFarmID *int64 `json:"defaultFarmId,omitempty" tf:"default_farm_id,omitempty"`
 
 	// Default ssl served to your customer
-	DefaultSSLID *float64 `json:"defaultSslId,omitempty" tf:"default_ssl_id,omitempty"`
+	DefaultSSLID *int64 `json:"defaultSslId,omitempty" tf:"default_ssl_id,omitempty"`
 
 	// Deny IP Load Balancing access to these ip block. No restriction if null. You cannot specify both allowed_source and denied_source at the same time. List of IP blocks.
 	// +listType=set
@@ -108,11 +108,11 @@ type TCPFrontendParameters struct {
 
 	// Default TCP Farm of your frontend
 	// +kubebuilder:validation:Optional
-	DefaultFarmID *float64 `json:"defaultFarmId,omitempty" tf:"default_farm_id,omitempty"`
+	DefaultFarmID *int64 `json:"defaultFarmId,omitempty" tf:"default_farm_id,omitempty"`
 
 	// Default ssl served to your customer
 	// +kubebuilder:validation:Optional
-	DefaultSSLID *float64 `json:"defaultSslId,omitempty" tf:"default_ssl_id,omitempty"`
+	DefaultSSLID *int64 `json:"defaultSslId,omitempty" tf:"default_ssl_id,omitempty"`
 
 	// Deny IP Load Balancing access to these ip block. No restriction if null. You cannot specify both allowed_source and denied_source at the same time. List of IP blocks.
 	// +kubebuilder:validation:Optional
